+++ src/main/java/reseau/project/status/items/ItemInterface.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface ItemInterface {
     public List<Item> getAllItemsOfACatalog(String catalogId);
     public Item getItemByCatalogAndItemId(String catalogId, String itemId);
-    public Item addOneItemToCatalog(String catalogId, String itemId, String itemName,MultipartFile itemImage, String itemDescription, double itemPrice, int itemRating ) throws IOException;
+    public Item addOneItemToCatalog(String catalogId, String itemId, String itemName,MultipartFile itemImage, String itemDescription, double itemPrice, double itemRating ) throws IOException;
 
 
     Item deleteItemOfCatalog(String catalogId, String itemId);
